refactor: rbac 모듈을 domain/group/role로 이동 (FSD 정렬)
ROLE은 group 도메인에 강하게 묶인 개념이므로 shared/ 보다 domain/group/
아래가 적합. shared/rbac 폴더 제거 후 domain/group/role/index.ts로 이동.

import 6곳 일괄 갱신:
- domain/members/components/member-card.tsx
- domain/group/components/group-role-management.tsx
- features/cardset/components/{cardset-create-form, cardset-update-form}.tsx
- pages/{group-detail, group-manage}.tsx

기능 변경 없음, 위치만 이동.
